Auth Task

+++ app/Http/Controllers/SalonSessionController.php
@@ -1,7 +1,7 @@
 <?php
 
 namespace App\Http\Controllers;
-
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use App\Models\SalonSession;
 use App\Models\User;
@@ -59,6 +59,10 @@ class SalonSessionController extends Controller
 
     public function update(Request $request, string $id)
     {
+        if (! Gate::allows('update-session', SalonSession::all()->where('id', $id)->first())) {
+            return redirect('/error')->with('message', 'У вас нет разрешения на редактирование сеанса номер ' . $id);
+        }
+
         $validated = $request->validate([
             'client_id'     => 'required|integer',
             'beautician_id' => 'required|integer',
@@ -76,6 +80,10 @@ class SalonSessionController extends Controller
 
     public function destroy(string $id)
     {
+        if (! Gate::allows('destroy-session', SalonSession::all()->where('id', $id)->first())) {
+            return redirect('/error')->with('message', 'У вас нет разрешения на удаление сеанса номер ' . $id);
+        }
+
         SalonSession::destroy($id);
         return redirect('/session');
     }
